bfgame: Rename .S to .sx
Keeping extensions in case-sensitive manner is asking for issues
in compilation between UNIX and Windows systems.

Assembly which needs preprocessor parsing can have either .S or .sx
extension. Switch to the latter.

Extension discrepencies introduced on Windows were causing a refusal
to parse on Linux, because GCC requires additional oprions to
preprocess a file which extension doesn't indicate the need for parsing.