bflibrary: Changed the approach to colour tables
It was impossible to rewrite the tables properly without closing
them into structs. So this is what has been made.

Also, separate functions for handling the whole struct were creaded,
and they use partial functions for fade table and ghost table
when neccessary.

Functions are now also better commented, and show warnings and
errors on misuse.