bflibrary: Added INI parser
The original library had no utility functions to parse INI files, and
each game implemented a simple one function parser when neccessary.
But the simple implementation lead to issues as it could not recognize
many of standard features of INI format (causing bad parse if a user
modified the file by hand), and was a bit different for each game,
causing confusion. This set of simple functions solves these problems.