bflibrary: Draw the mouse on WScreen
Previously, advanced mouse was being hacked into existence after
the content of WScreen was copied to a surface.

This needs to be simplified, so that mouse is drawn on WScreen
regardless whether it's advanced or not.

Fortunately SDL offers us proper API to create a surface from
app-owned buffer; we should use this.

The patch makes mouse drawing more universal, but it requires
a strange workaround with lbDrawSurface. Re-creation of
this surface should be implemented properly in further patches.

With this fix, bflibrary surface API is now useable for apps
with WScreen control.