bflibrary: Re-create DrawSureface on WScreen change by the app
If the app is in control of WScreen, then it should be able
to change it at any time. That causes a potential issue within
bflibrary, where DrawSurface can't be created around the WScreen
buffer on mode change, and assume the buffer link stands.

This patch introduces a WSchreen check function, and DrawSurface
re-creation if the WScreen starts pointing to a new buffer.

Such handling of the DrawSurface allows to use SDL blitting
capabilities and seeing results in WScreen. It also eliminates
one buffer copy when putting DrawSurface on ScreenSurface.