bfsoundlib: Fix sample callback to compile with GNU C23
Within the original SW assembly, the cb_sample_queue_callback is
called directly without parameters, but is also used as end-of-sample
callback, which does require parameters.

To make this compile, we were using the no-param declaration allowed
by ANSI-C to reproduce the behavior of any or no arguments accepted.
But there is no real need to do that - we can make the declaration
more strict instead.

While the SW Port code is not aimed at GNU dialect of ISO C, there is
also no real reason to keep this incompatible.

Co-authored-by: mefistotelis <mefistotelis@gmail.com>