bflibrary: declare the standars const-size types
The autotools should ensure the const-size types are available; but
we've added a prefix to autotools header defines - and this includes
the type definitions which autotools might add.

So to ensure the const-size types are declared, we need a few more
lines than the include of proper header, as we may need to re-declare
the types without prefix.