bflibrary: Proper way for keeping rescaled sprite sizes
Previous attempt at keeping the scaling size was not very good.
This time we're ensuring sprites are scaled to exactly the
expected size, not by enlarging first pixel, but by adding one
pixel to several first rows of the sprite.

Previously I considered a flag to enable the new scaling, but that
is not how engines should work - app requests size, engine should
deliver. Not deliver something similar, but exactly what was asked.