bflibrary: Fix mouse warping mode
The mode was damaged for some time, returning strange relative
positions. Now it should work as good as the direct coords mode.

The SDL2 also has a build-in relative mode, which allows to
achieve similar results without warping mouse; but that is
also capturing the mouse, while auto reset mode allows to
escape with fast mouse move.